feat: set up Alembic migrations and align monetary types to INT
Add Alembic with initial migration capturing the full schema.
Configure env.py to read DB_URL from project config module.
Change moms field from float to int in Pydantic model (öre).
Store all monetary values consistently in öre (SEK × 100).

+++ routers/classes/classes.py
@@ -40,7 +40,7 @@ class OrderData(BaseModel):
     subtotal: int
     deliveryFee: int = Field(ge=0)
     total: int = Field(ge=1)
-    moms: float = Field(ge=0)
+    moms: int = Field(ge=0)
 
 
 class CheckoutRequest(BaseModel):
